fix: use array args when calling DeleteSubdomain
There were two possible ways to fix this: change the calls (what this
commit does) or change the deleteSubdomain function to use the rest
operator.

At first it looked like the spread operator was the correct change
because that makes only one line of code incorrect vs two, however it
was then realized that functions like `deleteSubdomain` adhere to an
implied interface where the first argument to the function is always a
list of command arguments, so this solution was picked instead.