dev: add last-error command, context, and modules
This change first adds a concept of a "context" which is passed to the
command executor, and "modules" which can act on this context at
initialization.

We initialize this context with an event emitter so that all modules can
listen for or emit events.

Since commands have their cli arguments passed as a single argument to
the command function, this makes it possible to also pass the context
object to commands as a second argument. This was used to enable two
things:

- sites command (as an example) emits an event on error
- ErrorModule holds onto ERROR_BUFFER_LIMIT errors
- `last-error` command (new) asks ErrorModule for errors

The new command `last-error` references ErrorAPI directly which isn't
ideal; instead modules should be able to provide commands, but this
requires a refactor in how commands are loaded.